Add files via upload
This repository contains a simple and clean implementation of the Caesar Cipher algorithm using Python. The Caesar Cipher is one of the oldest classical encryption techniques, based on shifting characters in the alphabet by a fixed number of positions. This project provides functions to easily encrypt and decrypt messages with any custom shift value.